requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
optimum[onnxruntime]>=1.12.0
tqdm>=4.65.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0